restructured search functionality - now search occurs in search grid component
