+++ src/components/base-app.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import {Route, Link} from "react-router-dom";
+import {Route} from "react-router-dom";
 import NavigationBar from "./navigation-bar/navigation-bar";
 import HomePage from "./home-page/home-page";
 import DetailsPage from "./details-page/details-page";
@@ -28,7 +28,7 @@ class BaseApp extends React.Component {
                         <Login/>
                     </Route>
                     <Route path="/register"></Route>
-                    <Route path="/search">
+                    <Route exact path={["/search", "/search/:term"]}>
                         <SearchGrid/>
                     </Route>
                 </div>
